Fix ICC build break.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@8158 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Universal/SetupBrowserDxe/IfrParse.c
@@ -1083,7 +1083,7 @@ ParseOpCodes (
       //
       // The formset OpCode contains ClassGuid
       //
-      FormSet->NumberOfClassGuid = ((EFI_IFR_FORM_SET *) OpCodeData)->Flags & 0x3;
+      FormSet->NumberOfClassGuid = (UINT8) (((EFI_IFR_FORM_SET *) OpCodeData)->Flags & 0x3);
       CopyMem (FormSet->ClassGuid, OpCodeData + sizeof (EFI_IFR_FORM_SET), FormSet->NumberOfClassGuid * sizeof (EFI_GUID));
       break;
